Added more legacy flags
