+++ packages/expo-cli/src/commands/run/ios/runIos.ts
@@ -183,7 +183,7 @@ async function openInSimulatorAsync({
       bundleIdentifier,
     });
     if (result.status === 0) {
-      await SimControl.ensureSimulatorAppRunningAsync({ udid: device.udid });
+      await Simulator.ensureSimulatorAppRunningAsync({ udid: device.udid });
       await Simulator.activateSimulatorWindowAsync();
     } else {
       throw new CommandError(
